transfered the functions inside the app.jsx to a js file (inside appComponentFunction folder)
