+++ src/App.jsx
@@ -6,6 +6,16 @@ import Searchc from "./components/Search-create.jsx";
 import Swal from "sweetalert2";
 import { ToastContainer, toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
+import {
+  getAlumni,
+  searchAlumni,
+  addBatchID,
+  deleteBatchID,
+  deleteAlumni,
+  showConfirmationDeleteAlumni,
+  notifySuccessDelete,
+  showErrorDeleteAlumni,
+} from './components/appComponentFunction/AlumniOperations.js';
 
 //import "./components/sweetalert2/dist/sweetalert2.css";
 
@@ -17,100 +27,32 @@ function App() {
   const [batchID, setBatchID] = useState([]);
   const [mergedName, setMergedName] = useState("");
 
-  const searchAlumni = (search) => {
-    setAlumni(originalAlumni);
-    setAlumni((alums) => {
-      return alums.filter((alumnus) => {
-        const mergedName = (
-          alumnus.firstName +
-          " " +
-          alumnus.lastName
-        ).toLowerCase();
-        return mergedName.includes(search.toLowerCase());
-      });
-    });
-
-    console.log(alumni);
+  const handleGetAlumni = () => {
+    getAlumni(setAlumni, setOriginalAlumni);
   };
 
-  const getAlumni = async () => {
-    try {
-      const response = await fetch("http://localhost:5134/api/Alumni");
-      const data = await response.json();
-      setAlumni(data);
-      setOriginalAlumni(data);
-
-      console.log(data);
-    } catch (error) {
-      toast.error(error);
-    }
+  const handleSearch = (search) => {
+    searchAlumni(alumni, originalAlumni, setAlumni, search);
   };
 
-  //This is for the check and unchecking of checbox
-  // Function to add a batch ID
-  const addBatchID = (newID) => {
-    setBatchID([...batchID, newID]);
-
-    console.log(batchID);
+  const handleAddBatchID = (newID) => {
+    addBatchID(batchID, setBatchID, newID);
   };
 
-  // Function to delete a specific ID
-  const deleteBatchID = (idToDelete) => {
-    const updatedBatchIDs = batchID.filter((id) => id !== idToDelete);
-    setBatchID(updatedBatchIDs);
+  const handleDeleteBatchID = (idToDelete) => {
+    deleteBatchID(idToDelete, setBatchID, batchID);
+  }
 
-    console.log(batchID);
+  const handleDeleteAlumni = () => {
+    deleteAlumni(batchID, setBatchID, handleGetAlumni, notifySuccessDelete);
   };
 
-  const deleteAlumni = async () => {
-    try {
-      const deleteRequests = batchID.map(async (id) => {
-        const response = await fetch(`http://localhost:5134/api/Alumni/${id}`, {
-          method: "DELETE",
-        });
-
-        if (!response.ok) {
-          // Handle error if needed
-          console.error(`Failed to delete alumnus with ID ${id}`);
-        }
-      });
-
-      // Wait for all delete requests to complete
-      await Promise.all(deleteRequests);
-
-      setBatchID([]); //clear the compiled batch IDs
-      getAlumni();
-      notifySuccessDelete();
-    } catch (error) {
-      console.log(error);
-    }
+  const handleShowConfirmationDeleteAlumni = () => {
+    showConfirmationDeleteAlumni(Swal, handleDeleteAlumni);
   };
 
-  const showConfirmationDeleteAlumni = () => {
-    Swal.fire({
-      title: "Are you sure?",
-      text: "You won't be able to revert this!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Yes, delete it!",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        //call the deletion
-        deleteAlumni();
-      }
-    });
-  };
-
-  const notifySuccessDelete = () => toast.success("Successfully deleted");
-
-  const showErrorDeleteAlumni = () => {
-    Swal.fire({
-      icon: "error",
-      title: "T^T",
-      text: "Select an alumnus to delete first.",
-    });
+  const handleShowErrorDeleteAlumni = () => {
+    showErrorDeleteAlumni(Swal);
   };
 
   return (
@@ -118,19 +60,19 @@ function App() {
       <ToastContainer />
       <NavigationBar />
       <Searchc
-        getAlumni={getAlumni}
-        getSearchValue={searchAlumni}
+        getAlumni={handleGetAlumni}
+        getSearchValue={handleSearch}
         deleteAlumni={
           batchID.length === 0
-            ? showErrorDeleteAlumni
-            : showConfirmationDeleteAlumni
+            ? handleShowErrorDeleteAlumni
+            : handleShowConfirmationDeleteAlumni
         }
       />
       <AccordionList
         data={alumni}
-        getAlumni={getAlumni}
-        addBatchID={addBatchID}
-        deleteBatchID={deleteBatchID}
+        getAlumni={handleGetAlumni}
+        addBatchID={handleAddBatchID}
+        deleteBatchID={handleDeleteBatchID}
       />
       <Footer />
     </>
